# Micro optimization: build Hash w/ `{}`